WIP: Baumanl/mult_kern_vsns (#19)
Intended to add support for other kernel versions. Was not completely successful, but I laid down a first attempt at generalizing the infrastructure. I first tried with v4.0 and that was a lot different than v5.18, so I thought I would try v5.2 but that actually seemed to have even more issues somehow. I left it set up so that it should work for v5.18.